requirements.txt
pandas
Flask-SQLAlchemy
swagger-parser
flask
PyYAML